Added generate_new_clone method to figures and refactored the draw method
A method generate_new_clone has been added to the abstract class 'Figure' and implemented in Zeta, Ele, and Square subclasses to facilitate the creation of new instances of the figures. This is part of the improvement on game's logic. Also, a draw_figure function was created in the main script to avoid code repetition. Some changes were made in Game's class: a score attribute was added as well as the attribute 'next_figure' to give a preview of the following figure to the player. We also increased the width of the Pygame screen to accommodate the score panel. The sound was also activated.