Add Line figure and implement in Generator class
A new figure class named 'Line' is introduced in the figures directory, extending the base Figure class. It includes initialisation and positional transformations for the 'Line' figure. In addition, the 'Line' figure has been incorporated into the list of figures generated in the Generator class. This extension brings more complexity and challenge to the game providing a variety of shapes to the users.